feat: [#15] Add UIImage Extension.swift
이미지 사이즈를 조정해주도록 하는 함수를 UIImage Extension을 이용하여 구현
